organised the structure
